- first phase of testing done (tested all menu items in frontend, but without any actions: add to cart, wishlist, product on sale etc...) - fixed various bugs and errors

+++ components/com_redshop/models/cart.php
@@ -122,16 +122,19 @@ class cartModelcart extends JModelLegacy
             $cart = $session->get('cart');
             if ($cart)
             {
-                $idx = (int)($cart['idx']);
-                for ($j = 0; $j < $idx; $j++)
+                if (isset($cart['idx']))
                 {
-                    if (count($deletedrs) > 0 && in_array($cart[$j]['product_id'], $deletedrs))
+                    $idx = (int)($cart['idx']);
+                    for ($j = 0; $j < $idx; $j++)
                     {
-                        $this->delete($j);
-                    }
-                    if (count($includedrs) > 0 && !in_array($cart[$j]['product_id'], $includedrs))
-                    {
-                        $this->delete($j);
+                        if (count($deletedrs) > 0 && in_array($cart[$j]['product_id'], $deletedrs))
+                        {
+                            $this->delete($j);
+                        }
+                        if (count($includedrs) > 0 && !in_array($cart[$j]['product_id'], $includedrs))
+                        {
+                            $this->delete($j);
+                        }
                     }
                 }
             }
